Refactored to hooks
