+++ app/javascript/article-factcheck-video/video/shared.ts
@@ -1,6 +1,4 @@
-interface IVideo {
+export interface IVideo {
   getTime: () => number;
   goToTime: (time: number) => void;
 }
-
-export { IVideo };
