+++ package.json
@@ -66,7 +66,7 @@
     "prop-types": "^15.6.0",
     "query-string": "^6.1.0",
     "react": "^16.10.2",
-    "react-apollo": "^2.5.5",
+    "react-apollo": "3.1.3",
     "react-beautiful-dnd": "^7.1.3",
     "react-dom": "^16.10.2",
     "react-dropzone": "^4.2.9",
